Written by: Stefan Sperling <stsp@openbsd.org>
Orage was picking up ical.h via -I/usr/local/include but linking to
an older version of libical which is shipping in the orage tarball.
The header and the library are not compatible, some enum constants
don't match up. So orage could not read ical files correctly which
prevented e.g. appointments from showing up in the calendar.
Force inclusion of the ical headers which ship with orage to fix this.

"sure" landry@