Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import gnome-development-monitor-0.

GNOME Development Monitor is a simple graphical tool to follow the the
development activity of projects hosted on the GNOME infrastructure.

By analysis of commit records (for the last N days) it is able to
generate and display the following informations:

* Planet GNOME.
* Aggregate statistics, like GNOME Commit digest
  (http://blogs.gnome.org/commitdigest/).
* For each project:
    * Changes to NEWS.
    * Changes to ChangeLog.
    * Easy viewing of the commit log.
    * New patches in bugzilla.


WARNING: this project is still very young and a WIP.

some comments and ok jasper@