Written by: Stuart Henderson <sthen@openbsd.org>
Update to 1.6.0.21, various bugs (including some crashes) fixed.
This also has a small change in CDR generation, it's been well tested
upstream but still this can be a touchy area to change, so it's
going in now so the first OpenBSD release with Asterisk 1.6
packages has the change already made.

ok ajacoutot@