Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
SECURITY: Ensure keyboard grab and unlock dialog exist after monitor removal
(CVE-2010-0414)

gnome-screensaver currently doesn't deal with monitors getting
removed properly.  If the unlock dialog is on the removed monitor
then the unlock dialog and its associated keyboard grab are not
moved to an existing monitor when the monitor removal is processed.

This means that users can gain access to the locked system by placing
the mouse pointer on an external monitor and then disconnect the
external monitor.

https://bugzilla.gnome.org/show_bug.cgi?id=609337

ok jasper@