Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Update to foomatic-db-engine-4.0.4.

The important part for us is that "it is not required any more that for
Foomatic to return a PPD file for a given printer/driver combo that
there is an XML file for both the printer and the driver. One XML file
(either printer or driver) is enough when this file describes the
printer/driver relation."
That means that printers that did not appear before (i.e. one could not
create a custom ppd) where they were marked as supported will now
appear and work fine.