Written by: Stuart Henderson <sthen@openbsd.org>
Import devel/orc, from Brad with a patch from me to drop ARM machines
back to the generic support rather than trying to use the asm backend
which appears to be broken upstream.

Orc is a library and set of tools for compiling and executing
very simple programs that operate on arrays of data.  The "language"
is a generic assembly language that represents many of the features
available in SIMD architectures, including saturated addition and
subtraction, and many arithmetic operations.