Written by: Robert Nagy <robert@openbsd.org>
Remove the previous diff to mini.h because that was totally wrong, i made
it up before having my morning coffee.
So it turns out that sometimes the context on openbsd can be 0. This will
lead to problems. I suspect that this is because of our fantastic pthread
library. With the following diff we just skip doing some JIT stuff if the
context is 0 because the thread probably exited before.
Anyways this way everything *seems* to work fine and almost all regression
tests are passing now except for 2 out of 380.
Pepole understanding pthreads are welcome to look at it and come up with
something that is not a tripe XXX hack.