Written by: Landry Breuil <landry@openbsd.org>
Import sympa 6.0.1:
SYMPA is an electronic mailing list manager. It is used to automate list
management functions such as subscription, moderation and management of
archives. SYMPA also manages sending of messages to the lists, and makes
it possible to reduce the load on the system. It also provides a web
management interface, and an archive browser.

With feedback from sthen@ and ajacoutot@, ok ajacoutot@