Written by: Robert Nagy <robert@openbsd.org>
Remove the handles.c patch because that's not needed anymore because
Zoltan Varga from the mono team helped me to debug the problem we
were having. So the correct problem description is that on OpenBSD,
ctx can be NULL if we are interrupting poll().
The upstream diff also fixes the issue where we are interrupting
unmanaged code.