Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import p5-AnyEvent-HTTP 1.44

This module implements a simple, stateless and non-blocking HTTP client.
It supports GET, POST and other request methods, cookies and more, all
on a very low level. It can follow redirects supports proxies and
automatically limits the number of connections to the values specified
in the RFC.