Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import p5-AnyEvent-CouchDB 1.17

AnyEvent::CouchDB is a non-blocking CouchDB client implemented on top of
the AnyEvent framework. Using this library will give you the ability to
run many CouchDB requests asynchronously, and it was intended to be used
within a Coro+AnyEvent environment. However, it can also be used
synchronously if you want.