Written by: Jonathan Gray <jsg@openbsd.org>
Mpc is a C library for the arithmetic of complex numbers with
arbitrarily high precision and correct rounding of the result.
It is built upon and follows the same principles as Mpfr.

Imported with a PKGNAME of libmpc so as not to conflict with audio/mpc.
Incorporating changes from alek@