Written by: Robert Nagy <robert@openbsd.org>
Fix kinfo_proc2 code and add support for more mono stuff using sysctl to
get process informations. Remove the mkbundle.cs patch and replace it
with one that uses configure.in so the libgc dependency ends up in the
pkg-config file. Move disabled features to configure.in from our Makefile