Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import the g_desktop_app_info_get_filename() function from upstream.

This is required by newer GTK. We cannot just update Glib because it now
requires threads which will break p5-Glib2 dependent apps as our perl is
not threaded... solution to this is rthreads or ... <insert solution
here and send it to aja@>. Yes this means we are stuck with this version
for now.