Written by: Matthias Kilian <kili@openbsd.org>
Remove the native_bootstrap pseudo flavor. It's dangerous, because
the APIs of GHCs libraries depend on the version of the bootstrapping
compiler (and probably on the output of pom(6) and the amount of
active vulcanos in iceland).

See http://hackage.haskell.org/trac/ghc/ticket/4012#comment:3 for
details.

THIS ALSO MEANS THAT IF YOUR BUILD FAILS, YOU WILL HAVE TO MAKE
CLEAN AND START FROM SCRATCH! NEVER EVER TRY TO RESTART THE BUILD
OR THE LIBRARY ABIS WILL CHANGE! (Sorry for yelling)

Bump PKGNAME-main, since people may already have built ghc with
native_bootstrap.