Written by: Landry Breuil <landry@openbsd.org>
Import apvlv 0.0.9.6, A PDF Viewer Like Vim.
Apvlv is a PDF Viewer with vim-like keybindings using poppler rendering
library (that means it won't render pdf that evince and epdfview already
don't render).
For example, <Ctrl-f> for next page, <Ctrl-b> to previous page,
'k','j','h','l' to scrolling a page up, down, left or right, and so on.
It support lots of other shortcuts explained in the provided pdf
documentation, shown when starting it without arguments.