Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Create a link from ${LOCALBASE}/bin/smbspool to
${PREFIX}/libexec/cups/backend/smb so that is we have samba installed it
is possible to print to Windows printers out of the box (if samba is not
installed, the smb backend will not appear so there's no harm in doing
this).
Use FAKE_FLAGS instead ov post-install mv danse.
Fix group of cups.conf for dbus.