Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import ptsans 2.0

PT Sans family was developed as a part of the project "Public Types of
Russian Federation". The fonts of this project have open user license
and can be freely distributed. The main aim of the project is to give
possibility to the peoples of Russia to read and write on their native
languages.

from vadim zhukov (MAINTAINER), with tweaks by me