Written by: Marc Espie <espie@openbsd.org>
revert, for lots of technical reasons.
multi-packages are used when you have to build things together, but want
to separate the packages.

In this case, you have a bundle of things that do not even need to be
built together, that do not even share a MASTER_SITES, and were just bundled
because "they're related to vim".

This is deeply wrong. These files should be a separate port and packages.
Heck, they should be a separate port and several packages. Probably one
per-language.