Written by: Matthias Kilian <kili@openbsd.org>
GIO is striving to provide a modern, easy-to-use VFS API that sits
at the right level in the library stack.  The goal is to overcome
the shortcomings of GnomeVFS and provide an API that is so good
that developers prefer it over raw POSIX calls.  Among other things
that means using GObject.  It also means not cloning the POSIX API,
but providing higher-level, document-centric interfaces. This package
provides Haskell bindings to gio.