Written by: Matthias Kilian <kili@openbsd.org>
This is a helper port, used for fetching lilypond test data for the
regression tests only, so people who don't want to run the lilypond
regression tests don't have to download the huge distfile.

It is not meant to build or package anything, nor is it wired to
print/Makefile.

Suggested and ok, espie@ (some days ago)