Written by: Jacob Meuser <jakemsr@openbsd.org>
sndio support.  note, there are two different audio API APIs in jdk1.6,
one, HAE, is not portable to APIs that don't use file descriptors as
their handle, so it's still using ossaudio.  the other, DirectAudio,
is more portable, and is the one that this commit is converting to
sndio.  DirectAudio is preferred by jdk, and if it's supported, it will
be used.  in other words, your java apps on OpenBSD will now be using
sndio (assuing you're using jdk1.6, sndio for jdk1.7 coming soon).
just wanted to explain why this still links with libossaudio ...

ok ajacoutot