Written by: Robert Nagy <robert@openbsd.org>
Bump UTHREAD_SP_OFFSET to the correct value on i386 after the layout change
in struct _machdep_state from libpthread.
The offset is hardcoded in boehm-gc because it's a private struct so we
cannot use offsetof(). This will probably not change soon again but I am
writing this as a note for people who will run into the same problem in
the future.
All the regression tests are passing now and mono builds fine too.
Thanks to kurt@ for pointers.