Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Use "WORKGROUP" as a default workgroup name. It is the default in Mac OS
as well as previous Windows versions. At least it gives you a chance to
browse and mount smb shares with gvfs (or any other application using
smb.conf to get the workgroup name) without the need to configure
anything.

input and ok sthen@ on a previous version of this diff