Written by: Mark Kettenis <kettenis@openbsd.org>
The old version of boost included in this port assumes that GCC 4.1 and later
always provide the new atomic builtins.  In reality those are only available
on hardware that provides the necessary atomic instructions.  PA-RISC doesn't
have a suitable set of atomic instructions so work around it on hppa.

ok espie@, bernd@, landry@