Written by: Matthias Kilian <kili@openbsd.org>
Use a new bootstrapping compiler for ghc, linked against a current
libc. While here, switch to new REVISION/WANTLIB scheme.

Problem noticed by ajacoutot@

ok espie@ sthen@

Please note that any Haskell library depending on ghc-6.12.3
(cabal-wise) needs to be rebuilt. This affects devel/haddock
and devel/hs-QuickCheck, which will be bumped in a minute, but
if you've some libraries not contained in the ports tree, be
sure to double-check with ghc-pkg check.