Written by: Stuart Henderson <sthen@openbsd.org>
update chora, taking MAINTAINER, and reimport under ports/www/horde/chora.
agreed with avsm@

Chora is a tool for viewing code repositories that are managed
using the CVS or Subversion source control systems. It aims to
provide a high degree of integration with CVS, SVN and the other
Horde web-based tools.