Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import gnomint-1.2.1.

gnoMint is a tool for an easy creation and management of Certification
Authorities. It allows a fancy visualization of all the pieces that
conform a CA: x509 certificates, CSRs, CRLs...

Currently, it allows the creation of CAs, CSRs and Certificates, and
export both public and private parts of them into PEM formatted files.
It manages the revocation of the created certificates, as well as the
creation of CRLs.
<...>

ok jasper@