Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import liboobs-2.30.1

Liboobs is a wrapping library to the System Tools Backends, it will
provide easy to access GObjects to system configuration details, like
users, groups and network interfaces, it will handle sessions with the
backend and data consistency too.

This is a WIP, hence not hooked to the build.
DO NOT try this at home or you will harm yourself, you have been warned.

ok jasper@