Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import gnome-system-tools-2.30.2

These tools are intended to simplify the tasks of configuring a Unix
system for workstations. They are not intended for configuring Unix
servers.
<...>

This is a WIP, hence not hooked to the build.
DO NOT try this at home or you will harm yourself, you have been warned.

ok jasper@