Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import py-xmldiff 0.6.10

xmldiff is a Python tool that figures out the differences between two
similar XML files, in the same way the diff utility does it.
Xmldiff can be used as a library or as a command line tool. It can work
either with XML files or DOM trees.

ok aja@