Written by: Stuart Henderson <sthen@openbsd.org>
Zero newly allocated space, to avoid (in some cases) passing an
unclean buffer to Perl_is_utf8_string which proceeds to strlen() it,
causing an illegal access with malloc_flags J as it runs past the
end of the 0xd0d0d0.

Note there are various other problems with YAML::Syck and upstream
recommend not using it - see the README file.

ok jasper@