Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import frogatto-1.0.3

Frogatto is an open-source "platformer" or "jump-and-run" videogame.
Like in many classic games, the world is viewed as a cross-section seen
from the side, and your character (in this case, a small green fellow
named Frogatto) walks and jumps between solid platforms, whilst
avoiding monsters.

Based on the frogatto FreeBSD port.

ok sthen@