Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Make BROKEN, this is not compatible with the new version of vala I am
about to commit. Version 0.4 should fix this, but is still in
development and several of us need the new vala version to move ahead
with some ports.
FYI, nothing depends on this port in-tree.

ok jasper@