Written by: Jeremy Evans <jeremy@openbsd.org>
Bump all ruby ports not already handled.

In the upgrade from ruby 1.8.6 to 1.8.7, the PLISTs changed
due to differences in how RDoc processes files.

This also has a number of changes to the regress tests to
work with the changes to devel/ruby-rake.  It moves most of
the regress tests to use MODRUBY_REGRESS.

OK jcs@, landry@, jasper@, sthen@