Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Update to py-gtk2-2.17.0.
This was released as a development version a long time ago but we need
it as it contains the defines of our actual gtk+2 version. Other *nix do
the same.
We cannot update to the latest py-gtk stable version just yet as it
requires a major gtk+2 version update which I'll work on during+after
p2k10.

ok jasper@