Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import ruby-tilt 1.1

Tilt is a thin interface over a bunch of different Ruby template engines
in an attempt to make their usage as generic possible. This is useful
for web frameworks, static site generators, and other systems that
support multiple template engines but don't want to code for each of
them individually.

ok jeremy@