Written by: Jeremy Evans <jeremy@openbsd.org>
This adds an ugly hack to make gem install --user-install work
for gems with native extensions. Without this, it calls
/usr/bin/install -o root -g bin as a regular user, which fails due
to permission issues. This removes the -o root -g bin, so it can
succeed as a regular user.

This same patch was recently added to devel/ruby-gems, which is
specific to ruby 1.8.  ruby 1.9 ships with ruby-gems, so the
patch needs to be included here as well.

OK landry@