Written by: Jeremy Evans <jeremy@openbsd.org>
Add MODRUBY_RUBY_ADJ and MODRUBY_ADJ_FILES.  MODRUBY_RUBY_ADJ is a
perl command that accepts filename arguments and modifies the common
/usr/bin/env ruby shebang to ${RUBY}.  MODRUBY_ADJ_FILES is a short
cut that allows you to set filename patterns and have
MODRUBY_RUBY_ADJ called on all files in ${WRKSRC} that match that
pattern.  It adds a pre-configure action to do so, if a pre-configure
action is not already defined.  If a pre-configure action is already
defined, you can call the replacement command with
${MODRUBY_ADJ_REPLACE}.

ok sthen, landry