Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import vgrind from base

vgrind formats the program sources which are arguments in a nice style
using troff(1). Comments are placed in italics, keywords in bold face,
and the name of the current function is listed down the margin of each
page as it is encountered.

help and OK sthen@