Written by: Stephan A. Rickauer <stephan@openbsd.org>
import drupal6-lightbox2-1.11

The Lightbox2 module is a simple, unobtrusive script used to overlay
images on the current page. The module places images above your
current page, not within. This frees you from the constraints of the
layout, particularly column widths. It keeps users on the same page.

ok landry@