Written by: Steven Mestdagh <steven@openbsd.org>
import keepassx 0.4.3

KeePassX is an application for people with extremly high demands on
secure personal data management. It has a light interface, is cross
platform and published under the terms of the GNU General Public
License.

KeePassX saves many different information e.g. user names, passwords,
URLs, attachments and comments in one single database. For a better
management user-defined titles and icons can be specified for each
single entry. Furthermore the entries are sorted in groups, which are
customizable as well. The integrated search function allows to search
in a single group or the complete database.
KeePassX offers a little utility for secure password generation. The
password generator is very customizable, fast and easy to use.
Especially someone who generates passwords frequently will appreciate
this feature.

The complete database is always encrypted either with AES (alias
Rijndael) or Twofish encryption algorithm using a 256 bit key.
Therefore the saved information can be considered as quite safe.
KeePassX uses a database format that is compatible with KeePass
Password Safe. This makes the use of that application even more
favourable.

ok landry@