Written by: Pierre-Emmanuel Andre <pea@openbsd.org>
Import of vnstat:
vnStat is a console-based network traffic monitor for Linux and BSD that
keeps a log of network traffic for the selected interface(s). It uses the
network interface statistics provided by the kernel as information source.
This means that vnStat won't actually be sniffing any traffic and also
ensures light use of system resources.

With tweaks and advices from jasper@. Thanks

ok jasper@